+++ packages/runtime-wasi/src/syscalls/path.test.ts
@@ -199,6 +199,7 @@ describe('path_create_directory', () => {
       },
       rmSync: () => {},
       statSync: () => ({ isFile: false, isDirectory: true, size: 0, mtime: 0 }),
+      utimes: () => {},
     });
     const rc = t.ns.path_create_directory(3, 100, len);
     expect(rc).toBe(E_EXIST);
@@ -221,6 +222,7 @@ describe('path_create_directory', () => {
       },
       rmSync: () => {},
       statSync: () => ({ isFile: false, isDirectory: false, size: 0, mtime: 0 }),
+      utimes: () => {},
     });
     const rc = t.ns.path_create_directory(3, 100, len);
     expect(rc).toBe(E_ACCES);
@@ -243,6 +245,7 @@ describe('path_create_directory', () => {
       },
       rmSync: () => {},
       statSync: () => ({ isFile: false, isDirectory: false, size: 0, mtime: 0 }),
+      utimes: () => {},
     });
     const rc = t.ns.path_create_directory(3, 100, len);
     expect(rc).toBe(E_INVAL);
@@ -263,6 +266,7 @@ describe('path_create_directory', () => {
       },
       rmSync: () => {},
       statSync: () => ({ isFile: false, isDirectory: false, size: 0, mtime: 0 }),
+      utimes: () => {},
     });
     const rc = t.ns.path_create_directory(3, 100, len);
     expect(rc).toBe(E_NOENT);